requirements.txt
python-telegram-bot==13.7
azure-ai-formrecognizer==3.2.0
azure-cognitiveservices-speech==1.45.0
python-dotenv==0.19.2
requests==2.26.0